🧹 Move container & steps
Match Hyrax convention by using bulkrax/transactions.
